Refactor authentication and profile management: update password handling, improve error messages, and enhance user data retrieval

+++ learnify/src/components/core/Dashboard/Settings/DeleteAccount.jsx
@@ -9,12 +9,27 @@ export default function DeleteAccount() {
   const dispatch = useDispatch()
   const navigate = useNavigate()
 
+  
+  
+  
+
+
   async function handleDeleteAccount() {
-    try {
-      dispatch(deleteProfile(token, navigate))
-    } catch (error) {
-      console.log("ERROR MESSAGE - ", error.message)
+    const userConfirmed = window.confirm("Are you sure you want to proceed?");
+    if (userConfirmed) {
+      try {
+        dispatch(deleteProfile(token, navigate))
+      } catch (error) {
+        console.log("ERROR MESSAGE - ", error.message)
+      }
+      
+    
+    } else {
+      
+      
+     
     }
+   
   }
 
   return (
